+++ src/lib.rs
@@ -30,6 +30,7 @@
 mod date;
 mod error;
 mod month;
+mod raw;
 mod util;
 mod year;
 mod year_month;