+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+0.1.1:
+  * Fix `Date::add_days` for days in years 100, 200 and 300 of the 400 year cycle.
+
 0.1.0:
   * Initial release.